FEAT: Writing functional views
Change `polls/views.py` and add views `detail`, `results` and `vote`.
Adjust`polls/urls/py` accordingly. Change `idnex` view to display last
5 poll questions separated by commas and according to publication date.
Update `index` view in `polls/views.py` to use template. It loads the
template called `polls/index.html` and passes it a context. Use a
shortcut `render()` to fill a context and return an HttpResponse object.
The view (commented) raises the Http404 exception if a question with the
requested ID doesn't exist. We use a shortcut `get_object_or_404()` to
raise Http404 if the object doesn't exist. The `get_object_or_404()`
function takes a Django model as its first argument and an arbitrary
number of keyword arguments, which it passes to the get() function of
the model's manager. It raises Http404 if the object doesn't exist.