FEAT: Real implementation of the `vote()` function
Change vote function in the `polls/views.py` file.
`request.POST` is a dictionary-like object that lets you access
submitted data by key name. That means that `request.POST['choice']`
returns the ID of the selected choice, as a string (`request.POST`
values are always string). It will raise `KeyError` if choice wasn't
provided in POST data. The code checks for `KeyError` and redisplays
the question form with an error message if choice isn't given.
`F("votes") + 1` instructs the database to increase the voe count by 1.
F() objects assigned to model fields persist after saving the model
instance and will be applied on each save(). An F() object represents
the value of a model field, transformed value of a model field,
or annotated column. It makes it possible to refer to model field values
and perform database operations using them without actually having to
pull them out of the database into Python memory.
After incrementing the choice count, the code returns an
`HttpResponseRedirect` rather than a normal `HttpResponse`. It takes
a single argument: the URL to which the user will be redirected. We
should always return an HttpResponseRedirect after successfully dealing
wiht POST data.
We are using the `reverse()` function in the `HttpResponseRedirect`
constructor. This function helps avoid having hardcore a URL in the
view function. It is given the name of the view that we want to pass
control to and the variable portion of the URL pattern that points to
that view.