FEAT: Use generic views
Views `index()`, `detail()` and `results()` represent a common case
of basic web development: getting data from the DB according to a
parameter in the URL, loading a template and returning the rendered
template. For common patterns Django provides a shorcut, called the
"generic views" system. We use `ListView` to display a list of objects
and `DetailView` to display a detail page for a particular type of
object.
Change `polls/urls/py` and `polls/views.py`, delete/comment old views
and introduce new views based on Django's generic views.