FEAT: Customize the admin form
Change the admin options for a model. We create a model admin class and
pass it as the second argument to `admin.site.register()`. Add related
objects and customize the admin change list. Django offers a tabular way
of displaying inline related objects. We use `display()` decorator and
add filter to sidebar. In addition we add search capability that adds
a search box at the top of the change list.